Document API as not fully frozen.
This will give us the flexibility to continue to iterate and make
improvements to the API, instead of locking it down to whatever idea we
have first. This library is not yet used by that many clients, so minor
breaking changes are easy to update for.

See discussion in #3 for example of API decisions we need to make.